[update] move NewsEdit into NewsIndex
